% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyprint.R
\name{addEasyprint}
\alias{addEasyprint}
\title{Add easyPrint Plugin}
\usage{
addEasyprint(map, options = easyprintOptions())
}
\arguments{
\item{map}{a map widget object created from \code{\link[leaflet]{leaflet}}}

\item{options}{A named list of options. See \code{\link{easyprintOptions}}}
}
\value{
A leaflet map object
}
\description{
Add a control, which allows to print or export a map as .PNG.
}
\examples{
library(leaflet)
leaflet() \%>\%
  addTiles() \%>\%
  addEasyprint(options = easyprintOptions(
    title = "Print map",
    position = "bottomleft",
    exportOnly = TRUE
  ))

## Custom pixel size (e.g. high-res PNG in a Quarto/HTML document):
leaflet() \%>\%
  addTiles() \%>\%
  addEasyprint(options = easyprintOptions(
    title = "Save map to PNG",
    exportOnly = TRUE,
    filename = "map",
    tileWait = 1000,
    sizeModes = list(
      "CurrentSize",
      list(scale = 3, name = "3x current view")
    )
  ))
}
\references{
\url{https://github.com/rowanwins/leaflet-easyPrint}
}
\seealso{
Other EasyPrint Functions:
\code{\link[=easyprintMap]{easyprintMap()}},
\code{\link[=easyprintOptions]{easyprintOptions()}},
\code{\link[=removeEasyprint]{removeEasyprint()}}
}
\concept{EasyPrint Functions}
