% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contextmenu.R
\name{addItemContextmenu}
\alias{addItemContextmenu}
\title{addItemContextmenu}
\usage{
addItemContextmenu(map, option)
}
\arguments{
\item{map}{a map widget object created from \code{\link[leaflet]{leaflet}}}

\item{option}{new menu item to add}
}
\value{
A leaflet map object
}
\description{
Add a new contextmenu menu item
}
\seealso{
Other Contextmenu Functions:
\code{\link[=addContextmenu]{addContextmenu()}},
\code{\link[=context_mapmenuItems]{context_mapmenuItems()}},
\code{\link[=context_markermenuItems]{context_markermenuItems()}},
\code{\link[=context_menuItem]{context_menuItem()}},
\code{\link[=disableContextmenu]{disableContextmenu()}},
\code{\link[=enableContextmenu]{enableContextmenu()}},
\code{\link[=hideContextmenu]{hideContextmenu()}},
\code{\link[=insertItemContextmenu]{insertItemContextmenu()}},
\code{\link[=mapmenuItems]{mapmenuItems()}},
\code{\link[=markermenuItems]{markermenuItems()}},
\code{\link[=menuItem]{menuItem()}},
\code{\link[=removeItemContextmenu]{removeItemContextmenu()}},
\code{\link[=removeallItemsContextmenu]{removeallItemsContextmenu()}},
\code{\link[=setDisabledContextmenu]{setDisabledContextmenu()}},
\code{\link[=showContextmenu]{showContextmenu()}}
}
\concept{Contextmenu Functions}
