% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelgun.R
\name{addLabelgun}
\alias{addLabelgun}
\title{Add addLabelgun Plugin}
\usage{
addLabelgun(map, group, entries, weight)
}
\arguments{
\item{map}{A map widget object created from \code{\link[leaflet]{leaflet}}}

\item{group}{a group name}

\item{entries}{Higher value relates to faster insertion and slower search, and vice versa}

\item{weight}{the weight}
}
\value{
A leaflet map object
}
\description{
Add addLabelgun Plugin
}
\references{
\url{https://github.com/Geovation/labelgun}
}
