% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pruneCluster.R
\name{addPruneCluster}
\alias{addPruneCluster}
\title{addPruneCluster}
\usage{
addPruneCluster(
  map,
  lng = NULL,
  lat = NULL,
  layerId = NULL,
  group = NULL,
  icon = NULL,
  categoryField = NULL,
  color = NULL,
  popup = NULL,
  popupOptions = NULL,
  label = NULL,
  labelOptions = NULL,
  clusterOptions = NULL,
  clusterId = NULL,
  markerOptions = NULL,
  data = getMapData(map)
)
}
\arguments{
\item{map}{a map widget object created from \code{\link[leaflet]{leaflet}()}}

\item{lng}{a numeric vector of longitudes, or a one-sided formula of the form
\code{~x} where \code{x} is a variable in \code{data}; by default (if not
explicitly provided), it will be automatically inferred from \code{data} by
looking for a column named \code{lng}, \code{long}, or \code{longitude}
(case-insensitively)}

\item{lat}{a vector of latitudes or a formula (similar to the \code{lng}
argument; the names \code{lat} and \code{latitude} are used when guessing
the latitude column from \code{data})}

\item{layerId}{the layer id}

\item{group}{the name of the group the newly created layers should belong to
(for \code{\link[leaflet]{clearGroup}} and \code{\link[leaflet]{addLayersControl}} purposes).
Human-friendly group names are permitted--they need not be short,
identifier-style names. Any number of layers and even different types of
layers (e.g. markers and polygons) can share the same group name.}

\item{icon}{An icon or set of icons to include, created with \code{makeIcon} or \code{iconList}.}

\item{categoryField}{The column name used to categorize the charts.}

\item{color}{stroke color}

\item{popup}{The column name used to retrieve feature properties for the popup.}

\item{popupOptions}{A Vector of \code{\link[leaflet]{popupOptions}} to provide popups}

\item{label}{a character vector of the HTML content for the labels}

\item{labelOptions}{A Vector of \code{\link[leaflet]{labelOptions}} to provide label
options for each label. Default \code{NULL}}

\item{clusterOptions}{if not \code{NULL}, markers will be clustered using
\href{https://github.com/Leaflet/Leaflet.markercluster}{Leaflet.markercluster};
 you can use \code{\link[leaflet]{markerClusterOptions}()} to specify marker cluster
options}

\item{clusterId}{the id for the marker cluster layer}

\item{markerOptions}{Additional options for markers (see \code{\link[leaflet:markerOptions]{markerOptions::markerOptions()}}).}

\item{data}{the data object from which the argument values are derived; by
default, it is the \code{data} object provided to \code{leaflet()}
initially, but can be overridden}

\item{type}{The type of chart to use for clusters: \code{"pie"}, \code{"bar"}, \code{"horizontal"}, or \code{"custom"}.}

\item{popupFields}{A string or vector of strings indicating the column names to include in popups.}

\item{popupLabels}{A string or vector of strings indicating the labels for the popup fields.}

\item{options}{Additional options for cluster charts (see \code{\link{clusterchartOptions}}).}

\item{legendOptions}{A list of options for the legend, including the title and position.}
}
\description{
Adds cluster charts (pie, bar, horizontal, or custom) to a Leaflet map.
}
\details{
The `clusterCharts` use Leaflet's `L.DivIcon`, allowing you to fully customize
the styling of individual markers and clusters using CSS. Each individual marker
within a cluster is assigned the CSS class `clustermarker`, while the entire
cluster is assigned the class `clustermarker-cluster`. You can modify the appearance
of these elements by targeting these classes in your custom CSS.
}
\concept{pruneCluster}
