% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmsheader.R
\name{addWMSHeader}
\alias{addWMSHeader}
\title{Add WMS Layerwith custom Header}
\usage{
addWMSHeader(
  map,
  baseUrl,
  layerId = NULL,
  group = NULL,
  options = WMSTileOptions(),
  attribution = NULL,
  layers = NULL,
  header = NULL,
  data = getMapData(map)
)
}
\arguments{
\item{map}{a map widget object created from \code{\link[leaflet]{leaflet}()}}

\item{baseUrl}{a base URL of the WMS service}

\item{layerId}{the layer id}

\item{group}{the name of the group the newly created layers should belong to
(for \code{\link[leaflet]{clearGroup}} and \code{\link[leaflet]{addLayersControl}} purposes).
Human-friendly group names are permitted--they need not be short,
identifier-style names. Any number of layers and even different types of
layers (e.g. markers and polygons) can share the same group name.}

\item{options}{a list of extra options for tile layers, popups, paths
(circles, rectangles, polygons, ...), or other map elements}

\item{attribution}{the attribution text of the tile layer (HTML)}

\item{layers}{comma-separated list of WMS layers to show}

\item{data}{the data object from which the argument values are derived; by
default, it is the \code{data} object provided to \code{leaflet()}
initially, but can be overridden}
}
\value{
the new \code{map} object
}
\description{
Custom headers on Leaflet TileLayer WMS. It's a simple plugin that allow to
set custom header for WMS interface.
}
\references{
\url{https://github.com/ticinum-aerospace/leaflet-wms-header}
}
\seealso{
Other WMS Functions: 
\code{\link{addWMS}()}
}
\concept{WMS Functions}
