% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antpath.R
\name{clearAntpath}
\alias{clearAntpath}
\title{clearAntpath}
\usage{
clearAntpath(map)
}
\arguments{
\item{map}{a map widget object, possibly created from \code{\link[leaflet:leaflet]{leaflet()}}
but more likely from \code{\link[leaflet:leafletProxy]{leafletProxy()}}}
}
\value{
the new \code{map} object
}
\description{
Clear all Antpaths
}
\seealso{
Other Antpath Functions:
\code{\link[=addAntpath]{addAntpath()}},
\code{\link[=antpathOptions]{antpathOptions()}},
\code{\link[=removeAntpath]{removeAntpath()}}
}
\concept{Antpath Functions}
