% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layergroupconditional.R
\name{clearConditionalLayers}
\alias{clearConditionalLayers}
\title{clearConditionalLayers}
\usage{
clearConditionalLayers(map)
}
\arguments{
\item{map}{A map widget object created from \code{\link[leaflet]{leaflet}}.}
}
\description{
clearConditionalLayers
}
\seealso{
Other LayerGroupConditional Plugin:
\code{\link[=addLayerGroupConditional]{addLayerGroupConditional()}},
\code{\link[=removeConditionalLayer]{removeConditionalLayer()}}
}
\concept{LayerGroupConditional Plugin}
