% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexbin.R
\name{clearHexbin}
\alias{clearHexbin}
\title{clearHexbin}
\usage{
clearHexbin(map)
}
\arguments{
\item{map}{The map widget}
}
\value{
the new \code{map} object
}
\description{
Clears the data of the hexbinLayer.
}
\seealso{
Other Hexbin-D3 Functions:
\code{\link[=addHexbin]{addHexbin()}},
\code{\link[=hexbinOptions]{hexbinOptions()}},
\code{\link[=hideHexbin]{hideHexbin()}},
\code{\link[=showHexbin]{showHexbin()}},
\code{\link[=updateHexbin]{updateHexbin()}}
}
\concept{Hexbin-D3 Functions}
