% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history.R
\name{clearHistory}
\alias{clearHistory}
\title{clearHistory}
\usage{
clearHistory(map)
}
\arguments{
\item{map}{a map widget object created from
\code{\link[leaflet]{leafletProxy}}}
}
\value{
the new \code{map} object
}
\description{
Resets the stack of history items.
}
\references{
\url{https://github.com/cscott530/leaflet-history}
}
\seealso{
Other History Functions:
\code{\link[=addHistory]{addHistory()}},
\code{\link[=clearFuture]{clearFuture()}},
\code{\link[=goBackHistory]{goBackHistory()}},
\code{\link[=goForwardHistory]{goForwardHistory()}},
\code{\link[=historyOptions]{historyOptions()}}
}
\concept{History Functions}
