% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar.R
\name{closeSidebar}
\alias{closeSidebar}
\title{Close the Sidebar}
\usage{
closeSidebar(map, sidebar_id = NULL)
}
\arguments{
\item{map}{A leaflet map widget}

\item{sidebar_id}{The id of the sidebar (per \code{\link{sidebar_tabs}}).
Defaults to \code{NULL} such that the first sidebar is used.}
}
\value{
the new \code{map} object
}
\description{
Close the Sidebar
}
\seealso{
Other Sidebar Functions:
\code{\link[=addSidebar]{addSidebar()}},
\code{\link[=openSidebar]{openSidebar()}},
\code{\link[=removeSidebar]{removeSidebar()}},
\code{\link[=sidebar_pane]{sidebar_pane()}},
\code{\link[=sidebar_tabs]{sidebar_tabs()}}
}
\concept{Sidebar Functions}
