% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterCharts.R
\name{clusterchartOptions}
\alias{clusterchartOptions}
\title{clusterchartOptions}
\usage{
clusterchartOptions(
  rmax = 30,
  size = c(20, 20),
  width = 40,
  height = 50,
  strokeWidth = 1,
  innerRadius = 10,
  labelBackground = FALSE,
  labelFill = "white",
  labelStroke = "black",
  labelColor = "black",
  labelOpacity = 0.9,
  digits = 2,
  sortTitlebyCount = TRUE
)
}
\arguments{
\item{rmax}{The maximum radius of the clusters.}

\item{size}{The size of the cluster markers.}

\item{width}{The width of the bar-charts.}

\item{height}{The height of the bar-charts.}

\item{strokeWidth}{The stroke width of the chart.}

\item{innerRadius}{The inner radius of pie-charts.}

\item{labelBackground}{Should the label have a background? Default is `FALSE`}

\item{labelFill}{The label background color. Default is `white`}

\item{labelStroke}{The label stroke color. Default is `black`}

\item{labelColor}{The label color. Default is `black`}

\item{labelOpacity}{The label color. Default is `0.9`}

\item{digits}{The amount of digits. Default is `2`}

\item{sortTitlebyCount}{Should the svg-title be sorted by count or by the categories.}
}
\description{
Adds options for clusterCharts
}
\seealso{
Other clusterCharts:
\code{\link[=addClusterCharts]{addClusterCharts()}}
}
\concept{clusterCharts}
