% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contextmenu.R
\name{enableContextmenu}
\alias{enableContextmenu}
\title{enableContextmenu}
\usage{
enableContextmenu(map)
}
\arguments{
\item{map}{a map widget object created from \code{\link[leaflet]{leaflet}}}
}
\value{
A leaflet map object
}
\description{
Enable the contextmenu
}
\seealso{
Other Contextmenu Functions:
\code{\link[=addContextmenu]{addContextmenu()}},
\code{\link[=addItemContextmenu]{addItemContextmenu()}},
\code{\link[=context_mapmenuItems]{context_mapmenuItems()}},
\code{\link[=context_markermenuItems]{context_markermenuItems()}},
\code{\link[=context_menuItem]{context_menuItem()}},
\code{\link[=disableContextmenu]{disableContextmenu()}},
\code{\link[=hideContextmenu]{hideContextmenu()}},
\code{\link[=insertItemContextmenu]{insertItemContextmenu()}},
\code{\link[=mapmenuItems]{mapmenuItems()}},
\code{\link[=markermenuItems]{markermenuItems()}},
\code{\link[=menuItem]{menuItem()}},
\code{\link[=removeItemContextmenu]{removeItemContextmenu()}},
\code{\link[=removeallItemsContextmenu]{removeallItemsContextmenu()}},
\code{\link[=setDisabledContextmenu]{setDisabledContextmenu()}},
\code{\link[=showContextmenu]{showContextmenu()}}
}
\concept{Contextmenu Functions}
