% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geosearch.R
\name{geosearchProvider}
\alias{geosearchProvider}
\title{Provider for GeoSearch}
\usage{
geosearchProvider(
  type = c("OSM", "Bing", "Esri", "GeocodeEarth", "Google", "Here", "LocationIQ",
    "OpenCage", "Pelias", "Geoapify", "AMap", "GeoApiFr"),
  options = list()
)
}
\arguments{
\item{type}{The provider name}

\item{options}{Optional named list of options and parameters}
}
\value{
A list describing the provider config
}
\description{
Provider for GeoSearch
}
