% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{gibs_layers}
\alias{gibs_layers}
\title{The available GIBS layers with attributes}
\format{
An object of class \code{data.frame} with 276 rows and 4 columns.
}
\usage{
gibs_layers
}
\description{
The available GIBS layers with attributes
}
\keyword{datasets}
