% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexbin.R
\name{hideHexbin}
\alias{hideHexbin}
\title{hideHexbin}
\usage{
hideHexbin(map)
}
\arguments{
\item{map}{The map widget}
}
\value{
the new \code{map} object
}
\description{
Hide the hexbinLayer.
}
\seealso{
Other Hexbin-D3 Functions:
\code{\link[=addHexbin]{addHexbin()}},
\code{\link[=clearHexbin]{clearHexbin()}},
\code{\link[=hexbinOptions]{hexbinOptions()}},
\code{\link[=showHexbin]{showHexbin()}},
\code{\link[=updateHexbin]{updateHexbin()}}
}
\concept{Hexbin-D3 Functions}
