% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layergroupcollision.R
\name{LayerroupCollision}
\alias{LayerroupCollision}
\alias{addLayerGroupCollision}
\title{Add LayerroupCollision Plugin}
\usage{
addLayerGroupCollision(map, group = NULL, margin = 5)
}
\value{
A leaflet map object
}
\description{
Add LayerroupCollision Plugin
}
\references{
\url{https://github.com/Geovation/labelgun}
}
