% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet.extras2-package.R
\docType{package}
\name{leaflet.extras2-package}
\alias{leaflet.extras2}
\alias{leaflet.extras2-package}
\title{leaflet.extras2: Extra Functionality for 'leaflet' Package}
\description{
Several 'leaflet' plugins are integrated, which are available as extension to the 'leaflet' package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://trafficonese.github.io/leaflet.extras2/}
  \item \url{https://github.com/trafficonese/leaflet.extras2}
  \item Report bugs at \url{https://github.com/trafficonese/leaflet.extras2/issues}
}

}
\author{
\strong{Maintainer}: Gatscha Sebastian \email{sebastian_gatscha@gmx.at}

Authors:
\itemize{
  \item Gatscha Sebastian \email{sebastian_gatscha@gmx.at}
}

Other contributors:
\itemize{
  \item Ricardo Rodrigo Basa \email{radbasa@gmail.com} [contributor]
  \item Jeffrey O Hanson \email{jeffrey.hanson@uqconnect.edu.au} (\href{https://orcid.org/0000-0002-4716-6134}{ORCID}) [contributor]
}

}
\keyword{internal}
