% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leafletsync.R
\name{leafletsyncOptions}
\alias{leafletsyncOptions}
\title{leafletsync Options}
\usage{
leafletsyncOptions(
  noInitialSync = FALSE,
  syncCursor = TRUE,
  offsetFn = JS("function (center, zoom, refMap, tgtMap) { return center; }")
)
}
\arguments{
\item{noInitialSync}{Setting to \code{TRUE} disables initial synchronization
of the maps. The default is \code{FALSE}.}

\item{syncCursor}{The default \code{TRUE} adds a circle marker on the
synced map.}

\item{offsetFn}{A JavaScript-function to compute an offset for the center.}
}
\value{
A list of options for \code{addLeafletsync}
}
\description{
Additional list of options.
}
\seealso{
Other leafletsync Functions:
\code{\link[=addLeafletsync]{addLeafletsync()}},
\code{\link[=addLeafletsyncDependency]{addLeafletsyncDependency()}},
\code{\link[=isSynced]{isSynced()}},
\code{\link[=unsync]{unsync()}}
}
\concept{leafletsync Functions}
