% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contextmenu.R
\name{menuItem}
\alias{menuItem}
\title{menuItem}
\usage{
menuItem(text, callback = NULL, ...)
}
\arguments{
\item{text}{The label to use for the menu item}

\item{callback}{A callback function to be invoked when the menu item is
clicked. The callback is passed an object with properties identifying the
location the menu was opened at: \code{latlng}, \code{layerPoint}
and \code{containerPoint}. The callback-function must be valid JavaScript
and will be wrapped in \code{\link[leaflet]{JS}}.}

\item{...}{For further options please visit
\url{https://github.com/aratcliffe/Leaflet.contextmenu}}
}
\value{
A contextmenu item list
}
\description{
menuItem
}
\seealso{
Other Contextmenu Functions:
\code{\link[=addContextmenu]{addContextmenu()}},
\code{\link[=addItemContextmenu]{addItemContextmenu()}},
\code{\link[=context_mapmenuItems]{context_mapmenuItems()}},
\code{\link[=context_markermenuItems]{context_markermenuItems()}},
\code{\link[=context_menuItem]{context_menuItem()}},
\code{\link[=disableContextmenu]{disableContextmenu()}},
\code{\link[=enableContextmenu]{enableContextmenu()}},
\code{\link[=hideContextmenu]{hideContextmenu()}},
\code{\link[=insertItemContextmenu]{insertItemContextmenu()}},
\code{\link[=mapmenuItems]{mapmenuItems()}},
\code{\link[=markermenuItems]{markermenuItems()}},
\code{\link[=removeItemContextmenu]{removeItemContextmenu()}},
\code{\link[=removeallItemsContextmenu]{removeallItemsContextmenu()}},
\code{\link[=setDisabledContextmenu]{setDisabledContextmenu()}},
\code{\link[=showContextmenu]{showContextmenu()}}
}
\concept{Contextmenu Functions}
