% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar.R
\name{openSidebar}
\alias{openSidebar}
\title{Open the Sidebar by ID}
\usage{
openSidebar(map, id, sidebar_id = NULL, ns = NULL)
}
\arguments{
\item{map}{A leaflet map widget}

\item{id}{The id of the \code{\link{sidebar_pane}} to open.}

\item{sidebar_id}{The id of the sidebar (per \code{\link{sidebar_tabs}}).
Defaults to \code{NULL} such that the first sidebar is used.}

\item{ns}{The namespace function, if used in Shiny modules.}
}
\value{
the new \code{map} object
}
\description{
Open the Sidebar by ID
}
\seealso{
Other Sidebar Functions:
\code{\link[=addSidebar]{addSidebar()}},
\code{\link[=closeSidebar]{closeSidebar()}},
\code{\link[=removeSidebar]{removeSidebar()}},
\code{\link[=sidebar_pane]{sidebar_pane()}},
\code{\link[=sidebar_tabs]{sidebar_tabs()}}
}
\concept{Sidebar Functions}
