% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contextmenu.R
\name{removeallItemsContextmenu}
\alias{removeallItemsContextmenu}
\title{removeallItemsContextmenu}
\usage{
removeallItemsContextmenu(map)
}
\arguments{
\item{map}{a map widget object created from \code{\link[leaflet]{leaflet}}}
}
\value{
A leaflet map object
}
\description{
Remove all contextmenu items from the map.
}
\seealso{
Other Contextmenu Functions:
\code{\link[=addContextmenu]{addContextmenu()}},
\code{\link[=addItemContextmenu]{addItemContextmenu()}},
\code{\link[=context_mapmenuItems]{context_mapmenuItems()}},
\code{\link[=context_markermenuItems]{context_markermenuItems()}},
\code{\link[=context_menuItem]{context_menuItem()}},
\code{\link[=disableContextmenu]{disableContextmenu()}},
\code{\link[=enableContextmenu]{enableContextmenu()}},
\code{\link[=hideContextmenu]{hideContextmenu()}},
\code{\link[=insertItemContextmenu]{insertItemContextmenu()}},
\code{\link[=mapmenuItems]{mapmenuItems()}},
\code{\link[=markermenuItems]{markermenuItems()}},
\code{\link[=menuItem]{menuItem()}},
\code{\link[=removeItemContextmenu]{removeItemContextmenu()}},
\code{\link[=setDisabledContextmenu]{setDisabledContextmenu()}},
\code{\link[=showContextmenu]{showContextmenu()}}
}
\concept{Contextmenu Functions}
