% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrowhead.R
\name{removeArrowhead}
\alias{removeArrowhead}
\title{Remove arrowheads from Lines by layerId}
\usage{
removeArrowhead(map, layerId)
}
\arguments{
\item{map}{the map}

\item{layerId}{A single layerId or a vector of layerId's}
}
\value{
A modified leaflet map
}
\description{
Remove arrowheads from Lines by layerId
}
\seealso{
Other Arrowhead Functions:
\code{\link[=addArrowhead]{addArrowhead()}},
\code{\link[=arrowheadOptions]{arrowheadOptions()}},
\code{\link[=clearArrowhead]{clearArrowhead()}}
}
\concept{Arrowhead Functions}
