% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layergroupconditional.R
\name{removeConditionalLayer}
\alias{removeConditionalLayer}
\title{removeConditionalLayer}
\usage{
removeConditionalLayer(map, groups)
}
\arguments{
\item{map}{A map widget object created from \code{\link[leaflet]{leaflet}}.}

\item{groups}{A character vector of layer group names already added to the map.
These layer groups will be conditionally displayed based on the specified \code{conditions}.}
}
\description{
removeConditionalLayer
}
\seealso{
Other LayerGroupConditional Plugin:
\code{\link[=addLayerGroupConditional]{addLayerGroupConditional()}},
\code{\link[=clearConditionalLayers]{clearConditionalLayers()}}
}
\concept{LayerGroupConditional Plugin}
