% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyprint.R
\name{removeEasyprint}
\alias{removeEasyprint}
\title{removeEasyprint}
\usage{
removeEasyprint(map)
}
\arguments{
\item{map}{the map widget}
}
\value{
A leaflet map object
}
\description{
Removes the easyprint control from the map.
}
\seealso{
Other EasyPrint Functions:
\code{\link[=addEasyprint]{addEasyprint()}},
\code{\link[=easyprintMap]{easyprintMap()}},
\code{\link[=easyprintOptions]{easyprintOptions()}}
}
\concept{EasyPrint Functions}
