% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geosearch.R
\name{removeGeosearch}
\alias{removeGeosearch}
\title{removeGeosearch}
\usage{
removeGeosearch(map)
}
\arguments{
\item{map}{the map widget.}
}
\value{
the new \code{map} object
}
\description{
Remove the geosearch
}
\seealso{
Other Geosearch Functions:
\code{\link[=addGeosearch]{addGeosearch()}},
\code{\link[=geosearchOptions]{geosearchOptions()}}
}
\concept{Geosearch Functions}
