% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playback.R
\name{removePlayback}
\alias{removePlayback}
\title{removePlayback}
\usage{
removePlayback(map)
}
\arguments{
\item{map}{the map widget}
}
\value{
the new \code{map} object
}
\description{
Remove the Playback controls and markers.
}
\seealso{
Other Playback Functions:
\code{\link[=addPlayback]{addPlayback()}},
\code{\link[=playbackOptions]{playbackOptions()}}
}
\concept{Playback Functions}
