% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reachability.R
\name{removeReachability}
\alias{removeReachability}
\title{removeReachability}
\usage{
removeReachability(map)
}
\arguments{
\item{map}{the map widget.}
}
\value{
the new \code{map} object
}
\description{
Remove the reachability controls.
}
\seealso{
Other Reachability Functions:
\code{\link[=addReachability]{addReachability()}},
\code{\link[=reachabilityOptions]{reachabilityOptions()}}
}
\concept{Reachability Functions}
