% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar.R
\name{removeSidebar}
\alias{removeSidebar}
\title{Remove the Sidebar}
\usage{
removeSidebar(map, sidebar_id = NULL)
}
\arguments{
\item{map}{A leaflet map widget}

\item{sidebar_id}{The id of the sidebar (per \code{\link{sidebar_tabs}}).
Defaults to \code{NULL} such that the first sidebar is removed.}
}
\value{
the new \code{map} object
}
\description{
Remove the Sidebar
}
\seealso{
Other Sidebar Functions:
\code{\link[=addSidebar]{addSidebar()}},
\code{\link[=closeSidebar]{closeSidebar()}},
\code{\link[=openSidebar]{openSidebar()}},
\code{\link[=sidebar_pane]{sidebar_pane()}},
\code{\link[=sidebar_tabs]{sidebar_tabs()}}
}
\concept{Sidebar Functions}
