% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebyside.R
\name{removeSidebyside}
\alias{removeSidebyside}
\title{removeSidebyside}
\usage{
removeSidebyside(map, layerId = NULL)
}
\arguments{
\item{map}{a map widget}

\item{layerId}{the layer id of the \code{\link{addSidebyside}} layer}
}
\value{
the new \code{map} object
}
\description{
removeSidebyside
}
\seealso{
Other Sidebyside Functions:
\code{\link[=addSidebyside]{addSidebyside()}}
}
\concept{Sidebyside Functions}
