% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeslider.R
\name{removeTimeslider}
\alias{removeTimeslider}
\title{removeTimeslider}
\usage{
removeTimeslider(map)
}
\arguments{
\item{map}{the map widget}
}
\value{
the new \code{map} object
}
\description{
Remove the Timeslider controls and markers.
}
\seealso{
Other Timeslider Functions:
\code{\link[=addTimeslider]{addTimeslider()}},
\code{\link[=timesliderOptions]{timesliderOptions()}}
}
\concept{Timeslider Functions}
