% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorgrid.R
\name{removeVectorgrid}
\alias{removeVectorgrid}
\title{Remove a VectorGrid layer}
\usage{
removeVectorgrid(map, layerId)
}
\arguments{
\item{map}{a map widget object, possibly created from \code{\link[leaflet:leaflet]{leaflet()}}
but more likely from \code{\link[leaflet:leafletProxy]{leafletProxy()}}}

\item{layerId}{character vector; the layer id(s) of the item to remove}
}
\value{
the new \code{map} object
}
\description{
Remove a VectorGrid layer
}
\seealso{
Other Vectorgrid Functions:
\code{\link[=addProtobuf]{addProtobuf()}},
\code{\link[=addVectorgrid]{addVectorgrid()}},
\code{\link[=clearVectorgrid]{clearVectorgrid()}},
\code{\link[=vectorStyling]{vectorStyling()}}
}
\concept{Vectorgrid Functions}
