% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildings.R
\name{setBuildingData}
\alias{setBuildingData}
\title{Update the OSM-Buildings Data}
\usage{
setBuildingData(map, data)
}
\arguments{
\item{map}{A map widget object created from \code{\link[leaflet]{leaflet}}.}

\item{data}{A GeoJSON object containing Polygon features representing the buildings. The properties of these polygons can include attributes like \code{height}, \code{color}, \code{roofColor}, and others as specified in the OSM Buildings documentation.}
}
\description{
Update the OSM-Buildings Data
}
\seealso{
Other OSM-Buildings Plugin:
\code{\link[=addBuildings]{addBuildings()}},
\code{\link[=setBuildingStyle]{setBuildingStyle()}},
\code{\link[=updateBuildingTime]{updateBuildingTime()}}
}
\concept{OSM-Buildings Plugin}
