% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contextmenu.R
\name{setDisabledContextmenu}
\alias{setDisabledContextmenu}
\title{setDisabledContextmenu}
\usage{
setDisabledContextmenu(map, index, disabled = TRUE)
}
\arguments{
\item{map}{a map widget object created from \code{\link[leaflet]{leaflet}}}

\item{index}{Index of the contextmenu. (NOTE: Since the index is passed
to JavaScript, it is zero-based)}

\item{disabled}{Set to \code{TRUE} to disable the element and \code{FALSE}
to enable it. Default is \code{TRUE}}
}
\value{
A leaflet map object
}
\description{
Enable/Disable a contextmenu item by index.
}
\seealso{
Other Contextmenu Functions:
\code{\link[=addContextmenu]{addContextmenu()}},
\code{\link[=addItemContextmenu]{addItemContextmenu()}},
\code{\link[=context_mapmenuItems]{context_mapmenuItems()}},
\code{\link[=context_markermenuItems]{context_markermenuItems()}},
\code{\link[=context_menuItem]{context_menuItem()}},
\code{\link[=disableContextmenu]{disableContextmenu()}},
\code{\link[=enableContextmenu]{enableContextmenu()}},
\code{\link[=hideContextmenu]{hideContextmenu()}},
\code{\link[=insertItemContextmenu]{insertItemContextmenu()}},
\code{\link[=mapmenuItems]{mapmenuItems()}},
\code{\link[=markermenuItems]{markermenuItems()}},
\code{\link[=menuItem]{menuItem()}},
\code{\link[=removeItemContextmenu]{removeItemContextmenu()}},
\code{\link[=removeallItemsContextmenu]{removeallItemsContextmenu()}},
\code{\link[=showContextmenu]{showContextmenu()}}
}
\concept{Contextmenu Functions}
