% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velocity.R
\name{setOptionsVelocity}
\alias{setOptionsVelocity}
\title{setOptionsVelocity}
\usage{
setOptionsVelocity(map, layerId, options)
}
\arguments{
\item{map}{the map widget}

\item{layerId}{the layer id}

\item{options}{see \code{\link{velocityOptions}}}
}
\value{
the new \code{map} object
}
\description{
setOptionsVelocity
}
\seealso{
Other Velocity Functions:
\code{\link[=addVelocity]{addVelocity()}},
\code{\link[=removeVelocity]{removeVelocity()}},
\code{\link[=velocityOptions]{velocityOptions()}}
}
\concept{Velocity Functions}
