% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spin.R
\name{spinWhile}
\alias{spinWhile}
\title{Run work while a map spinner is visible (Shiny)}
\usage{
spinWhile(
  mapId,
  expr,
  options = NULL,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{mapId}{The \code{outputId} of the \code{leafletOutput}.}

\item{expr}{Expression to evaluate after the spinner is shown.}

\item{options}{Passed to \code{\link{startSpinner}}.}

\item{session}{The Shiny session; default is the current domain.}
}
\value{
The Shiny session (invisible).
}
\description{
Starts the spinner, waits until the browser has painted it (and Shiny
has flushed that update), then evaluates \code{expr} and always stops
the spinner.
}
\examples{
\dontrun{
observeEvent(input$go, {
  spinWhile("leaf", {
    Sys.sleep(2)
    leafletProxy("leaf") \%>\% addMarkers(data = quakes, ~long, ~lat)
  })
})
}
}
\seealso{
Other Spinner Functions:
\code{\link[=addSpinner]{addSpinner()}}
}
\concept{Spinner Functions}
