% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapkeyIcon.R
\name{[.leaflet_mapkey_icon_set}
\alias{[.leaflet_mapkey_icon_set}
\title{leaflet_mapkey_icon_set}
\usage{
\method{[}{leaflet_mapkey_icon_set}(x, i)
}
\arguments{
\item{x}{icons}

\item{i}{offset}
}
\description{
leaflet_mapkey_icon_set
}
\seealso{
Other Mapkey Functions:
\code{\link[=addMapkeyMarkers]{addMapkeyMarkers()}},
\code{\link[=makeMapkeyIcon]{makeMapkeyIcon()}},
\code{\link[=mapkeyIconList]{mapkeyIconList()}},
\code{\link[=mapkeyIcons]{mapkeyIcons()}}
}
\concept{Mapkey Functions}
