% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playback.R
\name{to_ms}
\alias{to_ms}
\title{to_ms
Change POSIX or Date to milliseconds}
\usage{
to_ms(data, time)
}
\arguments{
\item{data}{The data}

\item{time}{Columnname of the time column.}
}
\value{
A data.frame with the time column in milliseconds
}
\description{
to_ms
Change POSIX or Date to milliseconds
}
