% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorgrid.R
\name{vectorStyling}
\alias{vectorStyling}
\title{Default styles for OSM-based vector tiles}
\usage{
vectorStyling()
}
\value{
A named list of path options.
}
\description{
Layer-name styles for Mapbox, Nextzen, MapTiler / OpenMapTiles and similar
protobuf tiles. Override individual entries before passing to
\code{\link{addProtobuf}}.
}
\references{
\url{https://github.com/Leaflet/Leaflet.VectorGrid}
}
\seealso{
Other Vectorgrid Functions:
\code{\link[=addProtobuf]{addProtobuf()}},
\code{\link[=addVectorgrid]{addVectorgrid()}},
\code{\link[=clearVectorgrid]{clearVectorgrid()}},
\code{\link[=removeVectorgrid]{removeVectorgrid()}}
}
\concept{Vectorgrid Functions}
