% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildings.R
\name{addBuildings}
\alias{addBuildings}
\title{Add OSM-Buildings}
\usage{
addBuildings(
  map,
  apikey = NULL,
  layerId = NULL,
  group = NULL,
  opacity = 0.5,
  attribution = "© Map tiles <a href=\\"https://mapbox.com\\">Mapbox</a>"
)
}
\arguments{
\item{map}{A map widget object created from \code{\link[leaflet]{leaflet}}}

\item{options}{List of further options. See \code{\link{hexbinOptions}}}
}
\description{
Add OSM-Buildings
}
\note{
Out of the box a legend image is only available for Pressure,
  Precipitation Classic, Clouds Classic, Rain Classic, Snow, Temperature and
  Wind Speed.
}
\seealso{
https://osmbuildings.org/documentation/viewer/
}
\concept{OSM-Buildings Plugin}
