% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyprint.R
\name{addEasyprint}
\alias{addEasyprint}
\title{Add easyPrint Plugin}
\usage{
addEasyprint(map, options = easyprintOptions())
}
\arguments{
\item{map}{the map to add easyPrint to.}

\item{options}{A named list of options. See \code{\link{easyprintOptions}}}
}
\description{
Add easyPrint Plugin
}
\seealso{
\url{https://github.com/rowanwins/leaflet-easyPrint}

Other EasyPrint Plugin: 
\code{\link{easyprintMap}()},
\code{\link{easyprintOptions}()},
\code{\link{removeEasyprint}()}
}
\concept{EasyPrint Plugin}
