% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibs.R
\name{addGIBS}
\alias{addGIBS}
\title{Add GIBS Layers}
\usage{
addGIBS(
  map,
  layers = NULL,
  group = NULL,
  dates = NULL,
  opacity = 0.5,
  transparent = TRUE
)
}
\arguments{
\item{map}{A map widget object created from \code{\link[leaflet]{leaflet}}}

\item{layers}{vector of GIBS-layers. See \code{\link{gibs_layers}}}

\item{group}{the name of the group the newly created layers should belong to
(for\code{\link[leaflet]{clearGroup}} and
\code{\link[leaflet]{addLayersControl}} purposes). Human-friendly group
names are permitted–they need not be short, identifier-style names. Any
number of layers and even different types of layers (e.g. markers and
polygons) can share the same group name.}

\item{dates}{Date object. If multiple \code{layers} are added, you can add
a Date vector of the same length}

\item{opacity}{numeric value determining the opacity. If multiple
\code{layers} are added, you can add a numeric vector of the same length}

\item{transparent}{Should the layer be set `transparent`. If multiple
\code{layers} are added, you can add a boolean vector of the same length}
}
\description{
Add GIBS Layers
}
\seealso{
\url{https://github.com/aparshin/leaflet-GIBS}

Other GIBS Plugin: 
\code{\link{setDate}()},
\code{\link{setTransparent}()}
}
\concept{GIBS Plugin}
