% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexbin.R
\name{addHexbin}
\alias{addHexbin}
\title{Add a Hexbin layer}
\usage{
addHexbin(map, group = NULL, data = NULL)
}
\arguments{
\item{map}{a map widget object created from \code{\link[leaflet]{leaflet}}}
}
\description{
Add a Hexbin layer
}
