% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexbin.R
\name{addHexbin}
\alias{addHexbin}
\title{Add a Hexbin layer}
\usage{
addHexbin(
  map,
  lng = NULL,
  lat = NULL,
  radius = 20,
  layerId = NULL,
  group = NULL,
  opacity = 0.5,
  options = hexbinOptions(),
  data = getMapData(map)
)
}
\arguments{
\item{map}{A map widget object created from \code{\link[leaflet]{leaflet}}}

\item{lng}{a numeric vector of longitudes, or a one-sided formula of the form
~x where x is a variable in data; by default (if not explicitly provided),
it will be automatically inferred from data by looking for a column named
lng, long, or longitude (case-insensitively)}

\item{lat}{a vector of latitudes or a formula (similar to the lng argument;
the names lat and latitude are used when guessing the latitude column from
data)}

\item{radius}{Radius of the hexbin layer}

\item{layerId}{the layer id}

\item{group}{the name of the group the newly created layers should belong to
(for clearGroup and addLayersControl purposes). Human-friendly group names
are permitted–they need not be short, identifier-style names. Any number of
layers and even different types of layers (e.g. markers and polygons) can
share the same group name.}

\item{opacity}{Opacity on the hexbin layer}

\item{options}{List of further options. See \code{\link{hexbinOptions}}}

\item{data}{the data object from which the argument values are derived; by
default, it is the data object provided to leaflet() initially, but can be
overridden.}
}
\description{
Add a Hexbin layer
}
\note{
Currently doesn't respect \code{layerId} nor \code{group}
}
\seealso{
https://github.com/Asymmetrik/leaflet-d3#hexbins-api
}
