% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contextmenu.R
\name{addItemContextmenu}
\alias{addItemContextmenu}
\title{addItemContextmenu}
\usage{
addItemContextmenu(map, option)
}
\arguments{
\item{map}{a map widget object created from \code{\link[leaflet]{leaflet}}}

\item{option}{new menu item to add}
}
\value{
A leaflet map object
}
\description{
Add a new contextmenu menu item
}
\seealso{
Other Contextmenu Functions: 
\code{\link{addContextmenu}()},
\code{\link{context_mapmenuItems}()},
\code{\link{context_markermenuItems}()},
\code{\link{context_menuItem}()},
\code{\link{disableContextmenu}()},
\code{\link{enableContextmenu}()},
\code{\link{hideContextmenu}()},
\code{\link{insertItemContextmenu}()},
\code{\link{mapmenuItems}()},
\code{\link{markermenuItems}()},
\code{\link{menuItem}()},
\code{\link{removeItemContextmenu}()},
\code{\link{removeallItemsContextmenu}()},
\code{\link{setDisabledContextmenu}()},
\code{\link{showContextmenu}()}
}
\concept{Contextmenu Functions}
