% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leafletsync.R
\name{addLeafletsyncDependency}
\alias{addLeafletsyncDependency}
\title{Add the Leaflet Sync JS dependencies}
\usage{
addLeafletsyncDependency(map)
}
\arguments{
\item{map}{the map}
}
\value{
A modified leaflet map
}
\description{
Sometimes it makes sense to include the Leaflet Sync dependencies already
before synchronizing maps. For example, if you want to use the
`L.Sync.offsetHelper`. See the example in
\code{./inst/examples/offsetHelper.R}
}
\seealso{
Other leafletsync Functions: 
\code{\link{addLeafletsync}()},
\code{\link{isSynced}()},
\code{\link{leafletsyncOptions}()},
\code{\link{unsync}()}
}
\concept{leafletsync Functions}
