% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openweather.R
\name{addOpenweatherCurrent}
\alias{addOpenweatherCurrent}
\title{Add current OpenWeatherMap Marker}
\usage{
addOpenweatherCurrent(
  map,
  apikey = NULL,
  group = NULL,
  layerId = NULL,
  options = openweatherCurrentOptions()
)
}
\arguments{
\item{map}{A map widget object created from \code{\link[leaflet]{leaflet}}}

\item{apikey}{a valid Openweathermap-API key. Get one from
\href{https://openweathermap.org/api}{here}.}

\item{group}{name of the group}

\item{layerId}{the layer id}

\item{options}{List of further options. See
\code{\link{openweatherCurrentOptions}}}
}
\description{
Add current OpenWeatherMap Marker
}
\seealso{
https://github.com/trafficonese/leaflet-openweathermap

Other Openweather Plugin: 
\code{\link{addOpenweatherTiles}()},
\code{\link{openweatherCurrentOptions}()},
\code{\link{openweatherOptions}()}
}
\concept{Openweather Plugin}
