% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openweather.R
\name{addOpenweatherTiles}
\alias{addOpenweatherTiles}
\title{Add OpenWeather Layers}
\usage{
addOpenweatherTiles(
  map,
  apikey = NULL,
  layers = NULL,
  addControl = TRUE,
  group = NULL,
  opacity = 0.5
)
}
\arguments{
\item{map}{A map widget object created from \code{\link[leaflet]{leaflet}}}

\item{options}{List of further options. See \code{\link{hexbinOptions}}}
}
\description{
Add OpenWeather Layers
}
\note{
Out of the box a legend image is only available for Pressure,
  Precipitation Classic, Clouds Classic, Rain Classic, Snow, Temperature and
  Wind Speed.
}
\seealso{
https://github.com/Asymmetrik/leaflet-d3#hexbins-api

Other Openweather Plugin: 
\code{\link{addOpenweatherCurrent}()}
}
\concept{Openweather Plugin}
