% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playback.R
\name{addPlayback}
\alias{addPlayback}
\title{Add Playback to Leaflet}
\usage{
addPlayback(
  map,
  data,
  time = "time",
  icon = NULL,
  pathOptions = pathOptions(),
  options = playbackOptions()
)
}
\arguments{
\item{map}{a map widget}

\item{data}{data can either be a matrix or data.frame with coordinates, a
POINT Simple Feature or a list of POINT Simple Feature's. It can also be a
JSON string which must be in a specific form. See the Details for further
information.}

\item{time}{The column name of the time column. Default is \code{"time"}.}

\item{icon}{an icon which can be created with \code{\link[leaflet]{makeIcon}}}

\item{pathOptions}{style the CircleMarker with
\code{\link[leaflet]{pathOptions}}}

\item{options}{Lits of additional options. See \code{\link{playbackOptions}}}
}
\description{
Add Leaflet Playback Plugin based on the
 \href{https://github.com/hallahan/LeafletPlayback}{LeafletPlayback plugin}
}
\details{
If data is a JSON string, it must have the following form:
\preformatted{
{
  "type": "Feature",
  "geometry": {
    "type": "MultiPoint",
    "coordinates": [
      [-123.2653968, 44.54962188],
      [-123.26542599, 44.54951009]
    ]
},
  "properties": {
    "time": [1366067072000, 1366067074000]
  }
}
}
Additional arrays can be inside the properties, but are not required and are
not used by the plugin. If the JSON is stored in a file you can load it to R
via: \code{data <- paste(readLines(json_file_path, warn = F), collapse = "")}
}
\note{
If used in Shiny, you can listen to 2 events
\itemize{
 \item `map-ID`+"_pb_mouseover"
 \item `map-ID`+"_pb_click"
}
}
\seealso{
Other Playback Plugin: 
\code{\link{playbackOptions}()},
\code{\link{removePlayback}()}
}
\concept{Playback Plugin}
