% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorgrid.R
\name{addProtobuf}
\alias{addProtobuf}
\title{Add Vector Tiles through Protobuf}
\usage{
addProtobuf(
  map,
  urlTemplate = "//{s}.tile.openstreetmap.org/{z}/{x}/{y}.png",
  key = NULL,
  attribution = NULL,
  layerId = NULL,
  group = NULL,
  interactive = TRUE,
  options = leaflet::tileOptions(),
  popup = NULL,
  styling = vectorStyling(),
  data = getMapData(map)
)
}
\arguments{
\item{map}{a map widget object created from \code{\link[leaflet]{leaflet}()}}

\item{urlTemplate}{a character string as the URL template}

\item{key}{a valid API key}

\item{attribution}{the attribution text of the tile layer (HTML)}

\item{layerId}{the layer id}

\item{group}{the name of the group the newly created layers should belong to
(for \code{\link[leaflet]{clearGroup}} and \code{\link[leaflet]{addLayersControl}} purposes).
Human-friendly group names are permitted--they need not be short,
identifier-style names. Any number of layers and even different types of
layers (e.g. markers and polygons) can share the same group name.}

\item{interactive}{Should the layer fire mouse/pointer events. Default is TRUE}

\item{options}{a list of extra options for tile layers, popups, paths
(circles, rectangles, polygons, ...), or other map elements}

\item{popup}{TODO}

\item{styling}{TODO}

\item{data}{the data object from which the argument values are derived; by
default, it is the \code{data} object provided to \code{leaflet()}
initially, but can be overridden}
}
\value{
the new \code{map} object
}
\description{
A Leaflet plugin for working with Web Map services, providing:
}
\references{
\url{https://github.com/Leaflet/Leaflet.VectorGrid}
}
\seealso{
Other Vectorgrid Functions: 
\code{\link{addVectorgrid}()},
\code{\link{clearVectorgrid}()},
\code{\link{removeVectorgrid}()},
\code{\link{vectorStyling}()}
}
\concept{Vectorgrid Functions}
