% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar.R
\name{addSidebar}
\alias{addSidebar}
\title{Adds a Sidebar Leaflet Control}
\usage{
addSidebar(map, id = "sidebar", options = list(position = "left", fit = TRUE))
}
\arguments{
\item{map}{A leaflet map widget}

\item{id}{Id of the sidebar-div}

\item{options}{A named list with \code{position} and \code{fit} elements.}
}
\description{
Adds a Sidebar Leaflet Control
}
\seealso{
\url{https://github.com/Turbo87/sidebar-v2}

Other Sidebar Plugin: 
\code{\link{closeSidebar}()},
\code{\link{openSidebar}()},
\code{\link{removeSidebar}()},
\code{\link{sidebar_pane}()},
\code{\link{sidebar_tabs}()}
}
\concept{Sidebar Plugin}
