% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorgrid.R
\name{addVectorgrid}
\alias{addVectorgrid}
\title{Add a Vectorgrid}
\usage{
addVectorgrid(
  map,
  data,
  layerId = NULL,
  group = NULL,
  label = NULL,
  stroke = TRUE,
  color = "#03F",
  weight = 5,
  opacity = 0.5,
  fill = TRUE,
  fillColor = color,
  fillOpacity = 0.2,
  dashArray = NULL,
  smoothFactor = 1,
  noClip = FALSE,
  options = pathOptions()
)
}
\arguments{
\item{map}{a map widget object created from \code{\link[leaflet]{leaflet}()}}

\item{data}{the data object from which the argument values are derived; by
default, it is the \code{data} object provided to \code{leaflet()}
initially, but can be overridden}

\item{layerId}{the layer id}

\item{group}{the name of the group the newly created layers should belong to
(for \code{\link[leaflet]{clearGroup}} and \code{\link[leaflet]{addLayersControl}} purposes).
Human-friendly group names are permitted--they need not be short,
identifier-style names. Any number of layers and even different types of
layers (e.g. markers and polygons) can share the same group name.}

\item{label}{a character vector of the HTML content for the labels}

\item{stroke}{whether to draw stroke along the path (e.g. the borders of
polygons or circles)}

\item{color}{stroke color}

\item{weight}{stroke width in pixels}

\item{opacity}{stroke opacity (or layer opacity for tile layers)}

\item{fill}{whether to fill the path with color (e.g. filling on polygons or
circles)}

\item{fillColor}{fill color}

\item{fillOpacity}{fill opacity}

\item{dashArray}{a string that defines the stroke
\href{https://developer.mozilla.org/en/SVG/Attribute/stroke-dasharray}{dash
pattern}}

\item{smoothFactor}{how much to simplify the polyline on each zoom level
(more means better performance and less accurate representation)}

\item{noClip}{whether to disable polyline clipping}

\item{options}{a list of extra options for tile layers, popups, paths
(circles, rectangles, polygons, ...), or other map elements}
}
\value{
the new \code{map} object
}
\description{
A Leaflet plugin for working with Web Map services, providing:
}
\references{
\url{https://github.com/Leaflet/Leaflet.VectorGrid}
}
\seealso{
Other Vectorgrid Functions: 
\code{\link{addProtobuf}()},
\code{\link{clearVectorgrid}()},
\code{\link{removeVectorgrid}()},
\code{\link{vectorStyling}()}
}
\concept{Vectorgrid Functions}
