% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wms.R
\name{addWMS}
\alias{addWMS}
\title{Add Queryable WMS Layer}
\usage{
addWMS(
  map,
  baseUrl,
  layers = "",
  group = NULL,
  options = WMSTileOptions(),
  attribution = NULL,
  data = getMapData(map)
)
}
\arguments{
\item{map}{a map widget object created from \code{\link[leaflet]{leaflet}}}

\item{baseUrl}{a base URL of the WMS service}

\item{layers}{vector or list of WMS layers to show. The name of the layer is
used as the \code{layerId} (for \code{\link[leaflet]{removeTiles}}
purposes)}

\item{group}{the name of the group the newly created layers should belong to
(for clearGroup and addLayersControl purposes). Human-friendly group names
are permitted–they need not be short, identifier-style names. Any number of
layers and even different types of layers (e.g. markers and polygons) can
share the same group name. NOTE: If multiple \code{layers} are included,
the group argument will be ignored and the the layer-names will serve as
group names.}

\item{options}{List of further options. See
\code{\link[leaflet]{WMSTileOptions}}}

\item{attribution}{the attribution text of the tile layer (HTML)}

\item{data}{the data object from which the argument values are derived; by
default, it is the data object provided to \code{\link[leaflet]{leaflet}}
initially, but can be overridden.}
}
\description{
Add Queryable WMS Layer
}
\seealso{
https://github.com/heigeo/leaflet.wms
}
\concept{wms}
