% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antpath.R
\name{clearAntpath}
\alias{clearAntpath}
\title{Clear all Antpaths}
\usage{
clearAntpath(map)
}
\arguments{
\item{map}{a map widget object, possibly created from \code{\link[leaflet]{leaflet}}()
but more likely from \code{\link[leaflet]{leafletProxy}}()}
}
\description{
Clear all Antpaths
}
\seealso{
Other antpath: 
\code{\link{addAntpath}()},
\code{\link{antpathOptions}()},
\code{\link{removeAntpath}()}
}
\concept{antpath}
