% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrowhead.R
\name{clearArrowhead}
\alias{clearArrowhead}
\title{Remove arrowheads from Lines by group}
\usage{
clearArrowhead(map, group)
}
\arguments{
\item{map}{the map}

\item{group}{A group name}
}
\value{
A modified leaflet map
}
\description{
Remove arrowheads from Lines by group
}
\seealso{
Other Arrowhead Functions: 
\code{\link{addArrowhead}()},
\code{\link{arrowheadOptions}()},
\code{\link{removeArrowhead}()}
}
\concept{Arrowhead Functions}
