% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history.R
\name{clearFuture}
\alias{clearFuture}
\title{clearFuture
Resets the stack of future items.}
\usage{
clearFuture(map)
}
\arguments{
\item{map}{a map widget object created from \code{\link[leaflet]{leafletProxy}}}
}
\description{
clearFuture
Resets the stack of future items.
}
\examples{
library(leaflet)
leaflet()
}
\seealso{
\url{https://github.com/cscott530/leaflet-history}

Other History Functions: 
\code{\link{addHistory}()},
\code{\link{clearHistory}()},
\code{\link{goBackHistory}()},
\code{\link{goForwardHistory}()},
\code{\link{historyOptions}()}
}
\concept{History Functions}
