% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history.R
\name{clearFuture}
\alias{clearFuture}
\title{clearFuture}
\usage{
clearFuture(map)
}
\arguments{
\item{map}{a map widget object created from
\code{\link[leaflet]{leafletProxy}}}
}
\value{
the new \code{map} object
}
\description{
Resets the stack of future items.
}
\references{
\url{https://github.com/cscott530/leaflet-history}
}
\seealso{
Other History Functions: 
\code{\link{addHistory}()},
\code{\link{clearHistory}()},
\code{\link{goBackHistory}()},
\code{\link{goForwardHistory}()},
\code{\link{historyOptions}()}
}
\concept{History Functions}
