% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexbin.R
\name{clearHexbin}
\alias{clearHexbin}
\title{clearHexbin
Clear the hexbinLayer}
\usage{
clearHexbin(map)
}
\arguments{
\item{map}{The map widget}
}
\description{
clearHexbin
Clear the hexbinLayer
}
\seealso{
Other Hexbin-D3 Plugin: 
\code{\link{addHexbin}()},
\code{\link{hexbinOptions}()},
\code{\link{hideHexbin}()},
\code{\link{showHexbin}()},
\code{\link{updateHexbin}()}
}
\concept{Hexbin-D3 Plugin}
