% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history.R
\name{clearHistory}
\alias{clearHistory}
\title{clearHistory
Resets the stack of history items.}
\usage{
clearHistory(map)
}
\arguments{
\item{map}{a map widget object created from \code{\link[leaflet]{leafletProxy}}}
}
\description{
clearHistory
Resets the stack of history items.
}
\examples{
library(leaflet)
leaflet()
}
\seealso{
\url{https://github.com/cscott530/leaflet-history}

Other History Functions: 
\code{\link{addHistory}()},
\code{\link{clearFuture}()},
\code{\link{goBackHistory}()},
\code{\link{goForwardHistory}()},
\code{\link{historyOptions}()}
}
\concept{History Functions}
