% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorgrid.R
\name{clearVectorgrid}
\alias{clearVectorgrid}
\title{Clear all VectorGrid layers}
\usage{
clearVectorgrid(map)
}
\arguments{
\item{map}{a map widget object, possibly created from \code{\link[leaflet:leaflet]{leaflet()}}
but more likely from \code{\link[leaflet:leafletProxy]{leafletProxy()}}}
}
\value{
the new \code{map} object
}
\description{
Clear all VectorGrid layers
}
\seealso{
Other Vectorgrid Functions:
\code{\link[=addProtobuf]{addProtobuf()}},
\code{\link[=addVectorgrid]{addVectorgrid()}},
\code{\link[=removeVectorgrid]{removeVectorgrid()}},
\code{\link[=vectorStyling]{vectorStyling()}}
}
\concept{Vectorgrid Functions}
