% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorgrid.R
\name{clearVectorgrid}
\alias{clearVectorgrid}
\title{Clear all Vectorgrids}
\usage{
clearVectorgrid(map)
}
\arguments{
\item{map}{a map widget object, possibly created from \code{\link[leaflet]{leaflet}}()
but more likely from \code{\link[leaflet]{leafletProxy}}()}
}
\value{
the new \code{map} object
}
\description{
Clear all Vectorgrids
}
\seealso{
Other Vectorgrid Functions: 
\code{\link{addProtobuf}()},
\code{\link{addVectorgrid}()},
\code{\link{removeVectorgrid}()},
\code{\link{vectorStyling}()}
}
\concept{Vectorgrid Functions}
