% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar.R
\name{closeSidebar}
\alias{closeSidebar}
\title{Close the Sidebar}
\usage{
closeSidebar(map)
}
\arguments{
\item{map}{A leaflet map widget}
}
\description{
Close the Sidebar
}
\seealso{
Other Sidebar Plugin: 
\code{\link{addSidebar}()},
\code{\link{openSidebar}()},
\code{\link{removeSidebar}()},
\code{\link{sidebar_pane}()},
\code{\link{sidebar_tabs}()}
}
\concept{Sidebar Plugin}
