% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterCharts.R
\name{clusterchartOptions}
\alias{clusterchartOptions}
\title{clusterchartOptions}
\usage{
clusterchartOptions(
  rmax = 30,
  size = c(20, 20),
  width = 40,
  height = 50,
  strokeWidth = 1,
  pieMultiplier = 2,
  innerRadius = -10
)
}
\arguments{
\item{rmax}{The maxClusterRadius}

\item{size}{the size}

\item{width}{the width}

\item{height}{the height}

\item{strokeWidth}{the strokeWidth}

\item{pieMultiplier}{the pieMultiplier}

\item{innerRadius}{the innerRadius}
}
\description{
Adds options for clusterCharts
}
