% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterCharts.R
\name{clusterchartOptions}
\alias{clusterchartOptions}
\title{clusterchartOptions}
\usage{
clusterchartOptions(
  rmax = 30,
  size = c(20, 20),
  width = 40,
  height = 50,
  strokeWidth = 1,
  innerRadius = 10,
  labelBackground = FALSE,
  labelFill = "white",
  labelStroke = "black",
  labelColor = "black",
  labelOpacity = 0.9,
  aggregation = "sum",
  valueField = NULL,
  digits = 2,
  sortTitlebyCount = TRUE
)
}
\arguments{
\item{rmax}{The maximum radius of the clusters.}

\item{size}{The size of the cluster markers.}

\item{width}{The width of the bar-chart.}

\item{height}{The height of the bar-chart.}

\item{strokeWidth}{The stroke width in the chart.}

\item{innerRadius}{The inner radius of the pie-chart.}
}
\description{
Adds options for clusterCharts
}
