% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contextmenu.R
\name{context_mapmenuItems}
\alias{context_mapmenuItems}
\title{context_mapmenuItems}
\usage{
context_mapmenuItems(...)
}
\arguments{
\item{...}{contextmenu item/s}
}
\value{
A list of \code{context_menuItem} for the map
}
\description{
context_mapmenuItems
}
\seealso{
Other Contextmenu Functions: 
\code{\link{addContextmenu}()},
\code{\link{addItemContextmenu}()},
\code{\link{context_markermenuItems}()},
\code{\link{context_menuItem}()},
\code{\link{hideContextmenu}()},
\code{\link{insertItemContextmenu}()},
\code{\link{mapmenuItems}()},
\code{\link{markermenuItems}()},
\code{\link{menuItem}()},
\code{\link{removeItemContextmenu}()},
\code{\link{removeallItemsContextmenu}()},
\code{\link{setDisabledContextmenu}()},
\code{\link{showContextmenu}()}
}
\concept{Contextmenu Functions}
