% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contextmenu.R
\name{context_markermenuItems}
\alias{context_markermenuItems}
\title{context_markermenuItems}
\usage{
context_markermenuItems(...)
}
\arguments{
\item{...}{contextmenu item/s}
}
\value{
A list of \code{context_menuItem} for markers
}
\description{
context_markermenuItems
}
\seealso{
Other Contextmenu Functions: 
\code{\link{addContextmenu}()},
\code{\link{addItemContextmenu}()},
\code{\link{context_mapmenuItems}()},
\code{\link{context_menuItem}()},
\code{\link{disableContextmenu}()},
\code{\link{enableContextmenu}()},
\code{\link{hideContextmenu}()},
\code{\link{insertItemContextmenu}()},
\code{\link{mapmenuItems}()},
\code{\link{markermenuItems}()},
\code{\link{menuItem}()},
\code{\link{removeItemContextmenu}()},
\code{\link{removeallItemsContextmenu}()},
\code{\link{setDisabledContextmenu}()},
\code{\link{showContextmenu}()}
}
\concept{Contextmenu Functions}
