% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contextmenu.R
\name{context_menuItem}
\alias{context_menuItem}
\title{context_menuItem}
\usage{
context_menuItem(text, callback = NULL, ...)
}
\arguments{
\item{text}{The label to use for the menu item}

\item{callback}{A callback function to be invoked when the menu item is
clicked. The callback is passed an object with properties identifying the
location the menu was opened at: \code{latlng}, \code{layerPoint}
and \code{containerPoint}. The callback-function must be valid JavaScript
and will be wrapped in \code{\link[leaflet]{JS}}.}

\item{...}{For further options please visit
\url{https://github.com/aratcliffe/Leaflet.contextmenu}}
}
\value{
A contextmenu item list
}
\description{
context_menuItem
}
\seealso{
Other Contextmenu Functions: 
\code{\link{addContextmenu}()},
\code{\link{addItemContextmenu}()},
\code{\link{context_mapmenuItems}()},
\code{\link{context_markermenuItems}()},
\code{\link{hideContextmenu}()},
\code{\link{insertItemContextmenu}()},
\code{\link{mapmenuItems}()},
\code{\link{markermenuItems}()},
\code{\link{menuItem}()},
\code{\link{removeItemContextmenu}()},
\code{\link{removeallItemsContextmenu}()},
\code{\link{setDisabledContextmenu}()},
\code{\link{showContextmenu}()}
}
\concept{Contextmenu Functions}
