% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contextmenu.R
\name{enableContextmenu}
\alias{enableContextmenu}
\title{enableContextmenu}
\usage{
enableContextmenu(map)
}
\arguments{
\item{map}{a map widget object created from \code{\link[leaflet]{leaflet}}}
}
\value{
A leaflet map object
}
\description{
Enable the contextmenu
}
\seealso{
Other Contextmenu Functions: 
\code{\link{addContextmenu}()},
\code{\link{addItemContextmenu}()},
\code{\link{context_mapmenuItems}()},
\code{\link{context_markermenuItems}()},
\code{\link{context_menuItem}()},
\code{\link{disableContextmenu}()},
\code{\link{hideContextmenu}()},
\code{\link{insertItemContextmenu}()},
\code{\link{mapmenuItems}()},
\code{\link{markermenuItems}()},
\code{\link{menuItem}()},
\code{\link{removeItemContextmenu}()},
\code{\link{removeallItemsContextmenu}()},
\code{\link{setDisabledContextmenu}()},
\code{\link{showContextmenu}()}
}
\concept{Contextmenu Functions}
