% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history.R
\name{goBackHistory}
\alias{goBackHistory}
\title{goBackHistory
If possible, will go to previous map extent. Pushes current to the "future" stack.}
\usage{
goBackHistory(map)
}
\arguments{
\item{map}{a map widget object created from \code{\link[leaflet]{leafletProxy}}}
}
\description{
goBackHistory
If possible, will go to previous map extent. Pushes current to the "future" stack.
}
\examples{
library(leaflet)
leaflet()
}
\seealso{
\url{https://github.com/cscott530/leaflet-history}

Other History Functions: 
\code{\link{addHistory}()},
\code{\link{clearFuture}()},
\code{\link{clearHistory}()},
\code{\link{goForwardHistory}()},
\code{\link{historyOptions}()}
}
\concept{History Functions}
