% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history.R
\name{goBackHistory}
\alias{goBackHistory}
\title{goBackHistory}
\usage{
goBackHistory(map)
}
\arguments{
\item{map}{a map widget object created from
\code{\link[leaflet]{leafletProxy}}}
}
\value{
the new \code{map} object
}
\description{
If possible, will go to previous map extent. Pushes current extent to the
"future" stack.
}
\references{
\url{https://github.com/cscott530/leaflet-history}
}
\seealso{
Other History Functions: 
\code{\link{addHistory}()},
\code{\link{clearFuture}()},
\code{\link{clearHistory}()},
\code{\link{goForwardHistory}()},
\code{\link{historyOptions}()}
}
\concept{History Functions}
