% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history.R
\name{goForwardHistory}
\alias{goForwardHistory}
\title{goForwardHistory
If possible, will go to next map extent. Pushes current to the "back" stack.}
\usage{
goForwardHistory(map)
}
\arguments{
\item{map}{a map widget object created from \code{\link[leaflet]{leafletProxy}}}
}
\description{
goForwardHistory
If possible, will go to next map extent. Pushes current to the "back" stack.
}
\examples{
library(leaflet)
leaflet()
}
\seealso{
\url{https://github.com/cscott530/leaflet-history}

Other History Functions: 
\code{\link{addHistory}()},
\code{\link{clearFuture}()},
\code{\link{clearHistory}()},
\code{\link{goBackHistory}()},
\code{\link{historyOptions}()}
}
\concept{History Functions}
