% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contextmenu.R
\name{insertItemContextmenu}
\alias{insertItemContextmenu}
\title{insertItemContextmenu}
\usage{
insertItemContextmenu(map, option, index)
}
\arguments{
\item{map}{a map widget object created from \code{\link[leaflet]{leaflet}}}

\item{option}{new menu item to add}

\item{index}{Index of the contextmenu. (NOTE: Since the index is passed
to JavaScript, it is zero-based)}
}
\value{
A leaflet map object
}
\description{
Insert a new contextmenu menu item at a specific index
}
\seealso{
Other Contextmenu Functions: 
\code{\link{addContextmenu}()},
\code{\link{addItemContextmenu}()},
\code{\link{context_mapmenuItems}()},
\code{\link{context_markermenuItems}()},
\code{\link{context_menuItem}()},
\code{\link{disableContextmenu}()},
\code{\link{enableContextmenu}()},
\code{\link{hideContextmenu}()},
\code{\link{mapmenuItems}()},
\code{\link{markermenuItems}()},
\code{\link{menuItem}()},
\code{\link{removeItemContextmenu}()},
\code{\link{removeallItemsContextmenu}()},
\code{\link{setDisabledContextmenu}()},
\code{\link{showContextmenu}()}
}
\concept{Contextmenu Functions}
