% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contextmenu.R
\name{mapmenuItems}
\alias{mapmenuItems}
\title{mapmenuItems}
\usage{
mapmenuItems(...)
}
\arguments{
\item{...}{contextmenu item/s}
}
\value{
A list of \code{menuItem} for the map
}
\description{
mapmenuItems
}
\seealso{
Other Contextmenu Functions: 
\code{\link{addContextmenu}()},
\code{\link{addItemContextmenu}()},
\code{\link{hideContextmenu}()},
\code{\link{insertItemContextmenu}()},
\code{\link{markermenuItems}()},
\code{\link{menuItem}()},
\code{\link{removeItemContextmenu}()},
\code{\link{removeallItemsContextmenu}()},
\code{\link{setDisabledContextmenu}()},
\code{\link{showContextmenu}()}
}
\concept{Contextmenu Functions}
