% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contextmenu.R
\name{markermenuItems}
\alias{markermenuItems}
\title{markermenuItems}
\usage{
markermenuItems(...)
}
\arguments{
\item{...}{contextmenu item/s}
}
\value{
A list of \code{menuItem} for markers
}
\description{
markermenuItems
}
\seealso{
Other Contextmenu Functions: 
\code{\link{addContextmenu}()},
\code{\link{addItemContextmenu}()},
\code{\link{hideContextmenu}()},
\code{\link{insertItemContextmenu}()},
\code{\link{mapmenuItems}()},
\code{\link{menuItem}()},
\code{\link{removeItemContextmenu}()},
\code{\link{removeallItemsContextmenu}()},
\code{\link{setDisabledContextmenu}()},
\code{\link{showContextmenu}()}
}
\concept{Contextmenu Functions}
