% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movingmarker.R
\name{movingMarkerOptions}
\alias{movingMarkerOptions}
\title{Set options for Moving Markers}
\usage{
movingMarkerOptions(autostart = FALSE, loop = FALSE)
}
\arguments{
\item{autostart}{If \code{TRUE} the marker will start automatically after it
is added to map. Default is \code{FALSE}}

\item{loop}{if \code{TRUE} the marker will start automatically at the
beginning of the polyline when the it arrives at the end.
Default is \code{FALSE}}
}
\description{
Set options for Moving Markers
}
\references{
\url{https://github.com/ewoken/Leaflet.MovingMarker}
}
\seealso{
Other MovingMarker Functions: 
\code{\link{addMovingMarker}()},
\code{\link{startMoving}()}
}
\concept{MovingMarker Functions}
