% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar.R
\name{openSidebar}
\alias{openSidebar}
\title{Open the Sidebar by ID}
\usage{
openSidebar(map, id)
}
\arguments{
\item{map}{A leaflet map widget}

\item{id}{The id of the \code{\link{sidebar_pane}} to open}
}
\value{
the new \code{map} object
}
\description{
Open the Sidebar by ID
}
\seealso{
Other Sidebar Functions: 
\code{\link{addSidebar}()},
\code{\link{closeSidebar}()},
\code{\link{removeSidebar}()},
\code{\link{sidebar_pane}()},
\code{\link{sidebar_tabs}()}
}
\concept{Sidebar Functions}
