% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openweather.R
\name{openweatherOptions}
\alias{openweatherOptions}
\title{OpenWeatherMap Options}
\usage{
openweatherOptions(
  showLegend = TRUE,
  legendImagePath = NULL,
  legendPosition = c("bottomleft", "bottomright", "topleft", "topright")
)
}
\arguments{
\item{showLegend}{If \code{TRUE} and option \code{legendImagePath} is set
there will be a legend image on the map}

\item{legendImagePath}{A URL (is set to a default image for some layers, null
for others, see below). URL or relative path to an image which is a legend
to this layer}

\item{legendPosition}{Position of the legend images on the map. Must be one
of \code{'bottomleft', 'bottomright', 'topleft', 'topright'}}
}
\value{
A list of options for \code{addOpenweatherTiles}
}
\description{
OpenWeatherMap Options
}
\seealso{
Other Openweathermap Functions: 
\code{\link{addOpenweatherCurrent}()},
\code{\link{addOpenweatherTiles}()},
\code{\link{openweatherCurrentOptions}()}
}
\concept{Openweathermap Functions}
