% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openweather.R
\name{openweatherOptions}
\alias{openweatherOptions}
\title{OpenWeatherMap Options}
\usage{
openweatherOptions(
  showLegend = TRUE,
  legendImagePath = NULL,
  legendPosition = c("bottomleft", "bottomright", "topleft", "topright")
)
}
\arguments{
\item{showLegend}{If true and option 'legendImagePath' is set there will be a
legend image on the map.}

\item{legendImagePath}{URL (is set to a default image for some layers, null
for others, see below). URL or relative path to an image which is a legend
to this layer.}

\item{legendPosition}{Position of the legend images on the map. Available are
standard positions for Leaflet controls}
}
\description{
OpenWeatherMap Options
}
\seealso{
Other Openweather Plugin: 
\code{\link{addOpenweatherCurrent}()},
\code{\link{addOpenweatherTiles}()},
\code{\link{openweatherCurrentOptions}()}
}
\concept{Openweather Plugin}
