% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antpath.R
\name{removeAntpath}
\alias{removeAntpath}
\title{removeAntpath}
\usage{
removeAntpath(map, layerId = NULL)
}
\arguments{
\item{map}{a map widget object, possibly created from \code{\link[leaflet]{leaflet}}()
but more likely from \code{\link[leaflet]{leafletProxy}}()}

\item{layerId}{character vector; the layer id(s) of the item to remove}
}
\value{
the new \code{map} object
}
\description{
Remove one or more Antpaths from a map, identified by \code{layerId}.
}
\seealso{
Other Antpath Functions: 
\code{\link{addAntpath}()},
\code{\link{antpathOptions}()},
\code{\link{clearAntpath}()}
}
\concept{Antpath Functions}
