% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contextmenu.R
\name{removeItemContextmenu}
\alias{removeItemContextmenu}
\title{removeItemContextmenu}
\usage{
removeItemContextmenu(map, index)
}
\arguments{
\item{map}{a map widget object created from \code{\link[leaflet]{leaflet}}}

\item{index}{Index of the contextmenu. (NOTE: Since the index is passed to JavaScript,
it is zero-based)}
}
\value{
A leaflet map object
}
\description{
Remove a contextmenu item by index.
}
\seealso{
Other Contextmenu Functions: 
\code{\link{addContextmenu}()},
\code{\link{addItemContextmenu}()},
\code{\link{hideContextmenu}()},
\code{\link{insertItemContextmenu}()},
\code{\link{mapmenuItems}()},
\code{\link{markermenuItems}()},
\code{\link{menuItem}()},
\code{\link{removeallItemsContextmenu}()},
\code{\link{setDisabledContextmenu}()},
\code{\link{showContextmenu}()}
}
\concept{Contextmenu Functions}
