% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playback.R
\name{removePlayback}
\alias{removePlayback}
\title{removePlayback}
\usage{
removePlayback(map)
}
\arguments{
\item{map}{the map widget.}
}
\description{
Remove the reachability controls
}
\seealso{
Other Playback Plugin: 
\code{\link{addPlayback}()},
\code{\link{playbackOptions}()}
}
\concept{Playback Plugin}
