% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reachability.R
\name{removeReachability}
\alias{removeReachability}
\title{removeReachability}
\usage{
removeReachability(map)
}
\arguments{
\item{map}{the map widget.}
}
\description{
Remove the reachability controls
}
\seealso{
Other Reachability Functions: 
\code{\link{addReachability}()},
\code{\link{reachabilityOptions}()}
}
\concept{Reachability Functions}
