% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar.R
\name{removeSidebar}
\alias{removeSidebar}
\title{Remove the Sidebar}
\usage{
removeSidebar(map)
}
\arguments{
\item{map}{A leaflet map widget}
}
\value{
the new \code{map} object
}
\description{
Remove the Sidebar
}
\seealso{
Other Sidebar Functions: 
\code{\link{addSidebar}()},
\code{\link{closeSidebar}()},
\code{\link{openSidebar}()},
\code{\link{sidebar_pane}()},
\code{\link{sidebar_tabs}()}
}
\concept{Sidebar Functions}
