% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorgrid.R
\name{removeVectorgrid}
\alias{removeVectorgrid}
\title{Remove a Vectorgrid by layerId}
\usage{
removeVectorgrid(map, layerId)
}
\arguments{
\item{map}{a map widget object, possibly created from \code{\link[leaflet]{leaflet}}()
but more likely from \code{\link[leaflet]{leafletProxy}}()}

\item{layerId}{character vector; the layer id(s) of the item to remove}
}
\value{
the new \code{map} object
}
\description{
Remove a Vectorgrid by layerId
}
\seealso{
Other Vectorgrid Functions: 
\code{\link{addProtobuf}()},
\code{\link{addVectorgrid}()},
\code{\link{clearVectorgrid}()},
\code{\link{vectorStyling}()}
}
\concept{Vectorgrid Functions}
