% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velocity.R
\name{removeVelocity}
\alias{removeVelocity}
\title{removeVelocity}
\usage{
removeVelocity(map, group)
}
\arguments{
\item{map}{the map widget}

\item{group}{the group to remove}
}
\value{
the new \code{map} object
}
\description{
removeVelocity
}
\seealso{
Other Velocity Functions: 
\code{\link{addVelocity}()},
\code{\link{setOptionsVelocity}()},
\code{\link{velocityOptions}()}
}
\concept{Velocity Functions}
