% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibs.R
\name{setDate}
\alias{setDate}
\title{Set Date for GIBS Layers}
\usage{
setDate(map, layers = NULL, dates = NULL)
}
\arguments{
\item{map}{a map widget object created from \code{\link[leaflet:leaflet]{leaflet()}}}

\item{layers}{A character vector of GIBS-layers. See \code{\link{gibs_layers}}}

\item{dates}{Date object. If multiple \code{layers} are added, you can add a
Date vector of the same length}
}
\value{
the new \code{map} object
}
\description{
Set a new date for multi-temporal layers.
}
\seealso{
Other GIBS Functions: 
\code{\link{addGIBS}()},
\code{\link{setTransparent}()}
}
\concept{GIBS Functions}
