% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contextmenu.R
\name{setDisabledContextmenu}
\alias{setDisabledContextmenu}
\title{setDisabledContextmenu}
\usage{
setDisabledContextmenu(map, index, disabled = TRUE)
}
\arguments{
\item{map}{a map widget object created from \code{\link[leaflet]{leaflet}}}

\item{index}{Index of the contextmenu. (NOTE: Since the index is passed to JavaScript,
it is zero-based)}

\item{disabled}{Set to \code{TRUE} to disable the element and \code{FALSE}
to enable it. Default is \code{TRUE}}
}
\value{
A leaflet map object
}
\description{
Enable/Disable a contextmenu item by index.
}
\seealso{
Other Contextmenu Functions: 
\code{\link{addContextmenu}()},
\code{\link{addItemContextmenu}()},
\code{\link{context_mapmenuItems}()},
\code{\link{context_markermenuItems}()},
\code{\link{context_menuItem}()},
\code{\link{hideContextmenu}()},
\code{\link{insertItemContextmenu}()},
\code{\link{mapmenuItems}()},
\code{\link{markermenuItems}()},
\code{\link{menuItem}()},
\code{\link{removeItemContextmenu}()},
\code{\link{removeallItemsContextmenu}()},
\code{\link{showContextmenu}()}
}
\concept{Contextmenu Functions}
