% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velocity.R
\name{setOptionsVelocity}
\alias{setOptionsVelocity}
\title{setOptionsVelocity}
\usage{
setOptionsVelocity(map, layerId, options)
}
\arguments{
\item{map}{the map widget}

\item{layerId}{the layer id}

\item{options}{see \code{\link{velocityOptions}}}
}
\description{
setOptionsVelocity
}
\seealso{
Other Velocity Plugin: 
\code{\link{addVelocity}()},
\code{\link{removeVelocity}()},
\code{\link{velocityOptions}()}
}
\concept{Velocity Plugin}
