% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibs.R
\name{setTransparent}
\alias{setTransparent}
\title{Set Transparency for GIBS Layers}
\usage{
setTransparent(map, layers = NULL, transparent = TRUE)
}
\arguments{
\item{map}{A map widget object created from \code{\link[leaflet]{leaflet}}}

\item{layers}{vector of GIBS-layers. See \code{\link{gibs_layers}}}

\item{transparent}{Should the layer be transparent. If multiple
\code{layers} are added, you can add a boolean vector of the same length}
}
\description{
Set Transparency for GIBS Layers
}
\seealso{
Other GIBS Functions: 
\code{\link{addGIBS}()},
\code{\link{setDate}()}
}
\concept{GIBS Functions}
