% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibs.R
\name{setTransparent}
\alias{setTransparent}
\title{Set Transparency for GIBS Layers}
\usage{
setTransparent(map, layers = NULL, transparent = TRUE)
}
\arguments{
\item{map}{a map widget object created from \code{\link[leaflet:leaflet]{leaflet()}}}

\item{layers}{A character vector of GIBS-layers. See \code{\link{gibs_layers}}}

\item{transparent}{Should the layer be transparent. If multiple \code{layers}
are added, you can add a boolean vector of the same length}
}
\value{
the new \code{map} object
}
\description{
Change the transparency for no-data pixels.
}
\seealso{
Other GIBS Functions: 
\code{\link{addGIBS}()},
\code{\link{setDate}()}
}
\concept{GIBS Functions}
