% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wms.R
\name{setWMSParams}
\alias{setWMSParams}
\title{Update WMS request parameters}
\usage{
setWMSParams(map, layerId = NULL, group = NULL, ...)
}
\arguments{
\item{map}{a map widget object created from \code{\link[leaflet:leaflet]{leaflet()}}}

\item{layerId}{the layer id}

\item{group}{the name of the group the newly created layers should belong to
(for \code{\link[leaflet:clearGroup]{clearGroup()}} and \code{\link[leaflet:addLayersControl]{addLayersControl()}} purposes).
Human-friendly group names are permitted--they need not be short,
identifier-style names. Any number of layers and even different types of
layers (e.g., markers and polygons) can share the same group name.}

\item{...}{Named WMS request parameters to set, for example
\code{time = "2022-11-10T21:00:00Z"}.}
}
\value{
the new \code{map} object
}
\description{
Change WMS parameters such as \code{time}, \code{styles} or
\code{cql_filter} on an existing layer without rebuilding the map.
In Shiny, call this on \code{\link[leaflet]{leafletProxy}}.
}
\examples{
library(leaflet)
library(leaflet.extras2)

leaflet() \%>\%
  addTiles() \%>\%
  addWMS(
    baseUrl = "https://gibs.earthdata.nasa.gov/wms/epsg3857/best/wms.cgi",
    layers = "MODIS_Terra_CorrectedReflectance_TrueColor",
    layerId = "modis",
    options = WMSTileOptions(format = "image/jpeg", time = "2022-11-10")
  ) \%>\%
  setWMSParams(layerId = "modis", time = "2022-11-11")
}
\seealso{
Other WMS Functions:
\code{\link[=addWMS]{addWMS()}}
}
\concept{WMS Functions}
