% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexbin.R
\name{showHexbin}
\alias{showHexbin}
\title{showHexbin}
\usage{
showHexbin(map)
}
\arguments{
\item{map}{The map widget}
}
\value{
the new \code{map} object
}
\description{
Show the hexbinLayer.
}
\seealso{
Other Hexbin-D3 Functions: 
\code{\link{addHexbin}()},
\code{\link{clearHexbin}()},
\code{\link{hexbinOptions}()},
\code{\link{hideHexbin}()},
\code{\link{updateHexbin}()}
}
\concept{Hexbin-D3 Functions}
