% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexbin.R
\name{showHexbin}
\alias{showHexbin}
\title{showHexbin
Show the hexbinLayer}
\usage{
showHexbin(map)
}
\arguments{
\item{map}{The map widget}
}
\description{
showHexbin
Show the hexbinLayer
}
\seealso{
Other Hexbin-D3 Plugin: 
\code{\link{addHexbin}()},
\code{\link{clearHexbin}()},
\code{\link{hexbinOptions}()},
\code{\link{hideHexbin}()},
\code{\link{updateHexbin}()}
}
\concept{Hexbin-D3 Plugin}
