% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar.R
\name{sidebar_pane}
\alias{sidebar_pane}
\title{Create a Sidebar Panel}
\usage{
sidebar_pane(
  title = "Sidebar Title",
  id = NULL,
  icon = icon("caret-right"),
  ...
)
}
\arguments{
\item{title}{A title for the sidebar panel}

\item{id}{An id for the sidebar panel}

\item{icon}{An icon for the sidebar panel.}

\item{...}{List of elements to include in the panel}
}
\description{
Create a Sidebar Panel
}
\seealso{
Other sidebar-functions: 
\code{\link{addSidebar}()},
\code{\link{closeSidebar}()},
\code{\link{openSidebar}()},
\code{\link{removeSidebar}()},
\code{\link{sidebar_tabs}()}
}
\concept{sidebar-functions}
