% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playback.R
\name{to_jsonformat}
\alias{to_jsonformat}
\title{to_jsonformat
Transform object to JSON expected format}
\usage{
to_jsonformat(data, time)
}
\arguments{
\item{data}{The data}

\item{time}{Columnname of the time column.}
}
\value{
A list that is transformed to the expected JSON format
}
\description{
to_jsonformat
Transform object to JSON expected format
}
