% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leafletsync.R
\name{unsync}
\alias{unsync}
\title{Removes synchronization.}
\usage{
unsync(map, id = NULL, unsyncids = NULL)
}
\arguments{
\item{map}{the map}

\item{id}{The map id from which to unsynchronize the maps in \code{unsyncids}}

\item{unsyncids}{Unsynchronize the maps with the following IDs}
}
\value{
A map
}
\description{
Removes the synchronization of multiple maps from a specific map.
}
\seealso{
Other leafletsync Functions: 
\code{\link{addLeafletsync}()},
\code{\link{addLeafletsyncDependency}()},
\code{\link{isSynced}()},
\code{\link{leafletsyncOptions}()}
}
\concept{leafletsync Functions}
