% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildings.R
\name{updateBuildingTime}
\alias{updateBuildingTime}
\title{Update the Shadows OSM-Buildings with a POSIXct timestamp}
\usage{
updateBuildingTime(map, time)
}
\arguments{
\item{map}{A map widget object created from \code{\link[leaflet]{leaflet}}}
}
\description{
Update the Shadows OSM-Buildings with a POSIXct timestamp
}
\seealso{
https://osmbuildings.org/documentation/viewer/

Other OSM-Buildings Plugin: 
\code{\link{addBuildings}()}
}
\concept{OSM-Buildings Plugin}
