% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexbin.R
\name{updateHexbin}
\alias{updateHexbin}
\title{updateHexbin}
\usage{
updateHexbin(map, data = NULL, lng = NULL, lat = NULL, colorRange = NULL)
}
\arguments{
\item{map}{A map widget object created from \code{\link[leaflet]{leaflet}}}

\item{data}{the data object from which the argument values are derived; by
default, it is the data object provided to leaflet() initially, but can be
overridden.}

\item{lng}{a numeric vector of longitudes, or a one-sided formula of the form
~x where x is a variable in data; by default (if not explicitly provided),
it will be automatically inferred from data by looking for a column named
lng, long, or longitude (case-insensitively)}

\item{lat}{a vector of latitudes or a formula (similar to the lng argument;
the names lat and latitude are used when guessing the latitude column from
data)}

\item{colorRange}{Sets the range of the color scale used to fill the
hexbins on the layer.}
}
\description{
updateHexbin
}
\seealso{
Other Hexbin-D3 Functions: 
\code{\link{addHexbin}()},
\code{\link{clearHexbin}()},
\code{\link{hexbinOptions}()},
\code{\link{hideHexbin}()},
\code{\link{showHexbin}()}
}
\concept{Hexbin-D3 Functions}
