% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorgrid.R
\name{vectorStyling}
\alias{vectorStyling}
\title{vectorStyling}
\usage{
vectorStyling()
}
\value{
a list width different styles
}
\description{
Styling template for Mapbox/Nextzen/Maptiler and any OSM based vector ziles
}
\references{
\url{https://github.com/Leaflet/Leaflet.VectorGrid}
}
\seealso{
Other Vectorgrid Functions: 
\code{\link{addProtobuf}()},
\code{\link{addVectorgrid}()},
\code{\link{clearVectorgrid}()},
\code{\link{removeVectorgrid}()}
}
\concept{Vectorgrid Functions}
